+++ test/auth/auth_code_controller_test.dart
@@ -10,7 +10,7 @@ import 'dart:convert';
 
 void main() {
   Application<TestChannel> application;
-  TestClient client = new TestClient.onPort(8081);
+  TestClient client = new TestClient.onPort(8888);
 
   var codeResponse = (Map<String, String> form) {
     var m = new Map<String, String>.from(form);