+++ test/http/application_test.dart
@@ -42,13 +42,13 @@ void main() {
     });
 
     test("Application responds to request", () async {
-      var response = await http.get("http://localhost:8081/t");
+      var response = await http.get("http://localhost:8888/t");
       expect(response.statusCode, 200);
     });
 
     test("Application properly routes request", () async {
-      var tResponse = await http.get("http://localhost:8081/t");
-      var rResponse = await http.get("http://localhost:8081/r");
+      var tResponse = await http.get("http://localhost:8888/t");
+      var rResponse = await http.get("http://localhost:8888/r");
 
       expect(tResponse.body, '"t_ok"');
       expect(rResponse.body, '"r_ok"');
@@ -56,7 +56,7 @@ void main() {
 
     test("Application gzips content", () async {
       var resp = await http
-          .get("http://localhost:8081/t", headers: {"Accept-Encoding": "gzip"});
+          .get("http://localhost:8888/t", headers: {"Accept-Encoding": "gzip"});
       expect(resp.headers["content-encoding"], "gzip");
     });
 
@@ -65,7 +65,7 @@ void main() {
 
       var successful = false;
       try {
-        var _ = await http.get("http://localhost:8081/t");
+        var _ = await http.get("http://localhost:8888/t");
         successful = true;
       } catch (e) {
         expect(e, isNotNull);
@@ -73,7 +73,7 @@ void main() {
       expect(successful, false);
 
       await app.test();
-      var resp = await http.get("http://localhost:8081/t");
+      var resp = await http.get("http://localhost:8888/t");
       expect(resp.statusCode, 200);
     });
 
@@ -82,7 +82,7 @@ void main() {
         () async {
       var sum = 0;
       for (var i = 0; i < 10; i++) {
-        var result = await http.get("http://localhost:8081/startup");
+        var result = await http.get("http://localhost:8888/startup");
         sum += int.parse(JSON.decode(result.body));
       }
       expect(sum, 10);
@@ -113,7 +113,7 @@ void main() {
 
       crashingApp.options.context = {"crashIn": "dontCrash"};
       await crashingApp.test();
-      var response = await http.get("http://localhost:8081/t");
+      var response = await http.get("http://localhost:8888/t");
       expect(response.statusCode, 200);
       await crashingApp.stop();
     });