+++ test/http/ssl_test.dart
@@ -21,7 +21,7 @@ void main() {
       await app.start(numberOfInstances: 1);
 
       var completer = new Completer();
-      var socket = await SecureSocket.connect("localhost", 8081, onBadCertificate: (_) => true);
+      var socket = await SecureSocket.connect("localhost", 8888, onBadCertificate: (_) => true);
       var request = "GET /r HTTP/1.1\r\nConnection: close\r\nHost: localhost\r\n\r\n";
       socket.add(request.codeUnits);