+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 3.0.0
 
+- Change default port for `aqueduct serve` to 8888.
 - Binding metadata - `HTTPMethod`, `HTTPPath`, `HTTPBody`, `HTTPQuery` and `HTTPHeader` - have been changed to `Bind.method`, `Bind.path`, `Bind.body`, `Bind.query` and `Bind.header`, respectively.
 - Deprecates `@httpGet` and similar constants for `Bind.get()` and similar constructors.
 - Renames `Request.innerRequest` to `Request.raw`.